test: split oversized image router routes
